Finished implementing FilePath and moved tests into test sln. Added some basic types and includes into Monument project.
