+++ Core/GreatBath/FilePathProcessing.cpp
@@ -1,6 +1,6 @@
 #include "FilePathProcessing.h"
 
-#include "Core/Monument/Monument.h"
+#include "Core/Petra/basicmath.h"
 
 using namespace gbt;
 
@@ -14,14 +14,9 @@ FilePath::FilePath(const std::string& pathstring)
 	copy(pathstring);
 }
 
-FilePath::FilePath(FilePath&& newPath) noexcept
+FilePath::FilePath(c_string c_str)
 {
-	move(std::move(newPath));
-}
-
-FilePath::FilePath(const FilePath& oldPath)
-{
-	copy(oldPath);
+	*this = std::string(c_str);
 }
 
 FilePath& FilePath::operator=(const std::string& pathstring)
@@ -36,72 +31,32 @@ FilePath& FilePath::operator=(std::string&& pathstring) noexcept
 	return *this;
 }
 
-FilePath& FilePath::operator=(const FilePath& other)
+FilePath& FilePath::operator=(c_string c_str)
 {
-	copy(other);
-	return *this;
-}
-
-FilePath& FilePath::operator=(FilePath&& other) noexcept
-{
-	move(std::move(other));
+	*this = std::string(c_str);
 	return *this;
 }
 
 void FilePath::copy(const std::string& pathstring)
 {
 	m_path = pathstring;
-	generateViews();
+	generatePos();
 }
 
 void FilePath::move(std::string&& pathstring)
 {
 	m_path = std::move(pathstring);
-	generateViews();
+	generatePos();
 }
 
-void FilePath::copy(const FilePath& other)
+void FilePath::generatePos()
 {
-	m_path = other.m_path;
-	generateViews();
-}
-
-void FilePath::move(FilePath&& other)
-{
-	m_path = std::move(other.m_path);
-	m_folderPath = std::move(other.m_folderPath);
-	m_fileName = std::move(other.m_fileName);
-	m_fileNameNoExt = std::move(other.m_fileNameNoExt);
-	m_fileExt = std::move(other.m_fileExt);
-
-#ifdef _DEBUG
-	assert(m_folderPath.data() == nullptr || (m_folderPath.data() >= m_path.data() && m_folderPath.data() < m_path.data() + m_path.size()));
-	assert(m_fileName.data() == nullptr || (m_fileName.data() >= m_path.data() && m_fileName.data() < m_path.data() + m_path.size()));
-	assert(m_fileNameNoExt.data() == nullptr || (m_fileNameNoExt.data() >= m_path.data() && m_fileNameNoExt.data() < m_path.data() + m_path.size()));
-	assert(m_fileExt.data() == nullptr || (m_fileExt.data() >= m_path.data() && m_fileExt.data() < m_path.data() + m_path.size()));
-#endif
-}
-
-void FilePath::generateViews()
-{
-	const std::string_view pathView = std::string_view(m_path);
-
 	const size_t pathLen = m_path.length();
-	// TODO implement with min/max
-	size_t dotPos = pathView.rfind(FILE_EXT_DELIMITER);
-	const bool hadInvalidDot = dotPos == std::string_view::npos;
-	dotPos = hadInvalidDot ? pathLen : dotPos;
-	const size_t dotPosShifted = hadInvalidDot ? pathLen : dotPos + 1;
+	m_nameStartPos = m_path.find_last_of(FOLDER_DELIMITERS);
+	m_nameStartPos = m_nameStartPos == std::string::npos ? 0 : rqm::min( m_nameStartPos + 1, pathLen );
+	const size_t nameLen = pathLen - m_nameStartPos;
 
-	size_t slashPos = pathView.find_last_of(FOLDER_DELIMITERS);
-	const bool hadInvalidSlash = slashPos == std::string_view::npos;
-	const size_t slashPosShifted = hadInvalidDot ? pathLen : (hadInvalidSlash ? 0 : slashPos + 1);
-	slashPos = hadInvalidDot ? pathLen : (hadInvalidSlash ? 0 : slashPos);
-
-	const bool slashReasonable = slashPosShifted < pathLen;
-
-	m_folderPath = pathView.substr(0, slashPos);
-	m_fileName = slashReasonable ? pathView.substr(slashPosShifted) : std::string_view();
-	m_fileNameNoExt = slashReasonable ? pathView.substr(slashPosShifted, dotPos - slashPosShifted) : std::string_view();
-	m_fileExt = pathView.substr(dotPosShifted);
+	const std::string_view nameView = std::string_view(&(m_path[m_nameStartPos]), nameLen);
+	m_nameEndPos = rqm::min(nameView.rfind(FILE_EXT_DELIMITER), nameLen) + m_nameStartPos;
+	m_extStartPos = rqm::min(m_nameEndPos + 1, pathLen);
 }
