+++ Core/Monument/Monument.h
@@ -1,3 +1,4 @@
 #pragma once
 
 #include "assert.h"
+#include "basictypes.h"
