+++ Core/Monument/Monument.vcxproj
@@ -158,6 +158,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="assert.h" />
+    <ClInclude Include="basictypes.h" />
     <ClInclude Include="Monument.h" />
   </ItemGroup>
   <ItemGroup>
