+++ Core/Monument/Monument.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClInclude Include="Monument.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="basictypes.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="Monument.cpp">
