+++ Core/Monument/basictypes.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// null-terminated c-string, please do not use non null-terminated char* buffers if we're specifiying a c_string
+typedef const char* c_string;
