+++ Ctrl test/CtrlTest.vcxproj
@@ -156,7 +156,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="Source\CtrlTest.cpp" />
+    <ClCompile Include="CtrlTest.cpp" />
+    <ClCompile Include="GreatBathTests.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Core\GreatBath\GreatBath.vcxproj">
@@ -174,6 +175,9 @@
     <ProjectReference Include="..\Tools\SunDial Profiler\SunDial Profiler.vcxproj">
       <Project>{75cde69c-f1e7-4ce0-b20d-65e9c5a9dbe1}</Project>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="tests.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
